refactor(core): fix fmt/lint/build gates + Qwen 3.5+ native multimodal sniffing
Clean up the three code-quality gates that were failing on
(after Phase 43) and correct the Qwen strategy auto-sniff.

## lint (35 errors → 0)
- update_spend_log: allow too_many_arguments on the trait + 3 dialect impls
  + dispatch (14-arg signature already had precedent at db.rs:2180)
- db.rs PG date inlining: replace is_some()+unwrap() with if-let patterns
  (6 sites); map_or → is_some_and; needless borrow; allow on
  query_spend_logs_with_status_filter
- adapter.rs: option.map(|obj| …) unit closures → if let Some(obj);
  mem::take over drain(..).collect(); Default impls for both stream adapters
- engine.rs: allow too_many_arguments on 3 create_job dialect fns
- body_archive: derive+Default for cache config enums (was manual impl);
  += over = current +; needless &schema_desc borrows
- budget/resetter: EntityType now impls std::str::FromStr (callers migrated
  in budget/resetter + jobs.rs); Duration RangeInclusive::contains
- crypto.rs: consecutive str::replace('\n','')+replace('\r','') →
  replace(['\n','\r'],)
- router.rs: RouterStrategy impls FromStr; map_or → is_none_or
- aigw-server: 13 page_size .max(1).min(100) → .clamp(1,100); 12 redundant
  closures (usage.map(extract_*)); 6 redundant & in format!; save_result /
  calc_spend allow too_many_arguments; chat.rs/responses.rs collapsible
  if-in-match; spend.rs as_ref.map clone → clone; map().flatten() → and_then;
  very-complex-type split into ActivityMetadataRaw/ActivityRowRaw type aliases
- aigw-migrate: native.rs unnecessary usize casts; remote_import run_filtered
  allow + is_none_or
- main.rs: RequestIdMakeSpan::default() → unit struct

## fmt
- cargo fmt --all (21 files, style-only; matches rustfmt 1.8+ norms)

## Qwen strategy fix (infer_strategy)
- Qwen3.5 and ALL later versions (3.6/3.7/…/4.x/5.x…) are native unified
  multimodal (Qwen3_5ForConditionalGeneration embeds a vision encoder and
  top-level image/video token ids; no separate -VL suffix). Old check
  m.contains(qwen3) && vl||vision missed qwen3.5-* entirely.
- New is_qwen_native_multimodal() parses the version number:
  major>3 → true; qwen3.x with x>=5 → true; else false. qwen3 base and
  qwen2.x still require the -VL suffix.
- +1 UT (test_infer_strategy_qwen35_native_multimodal) covering 3.5/3.6/3.9/
  4.x/5.x multimodal + qwen3/qwen3.4 text-only boundary. 19 image_tokens UT.

## verification
- task fmt PASS, task lint PASS (0 errors), task build PASS (release binary)
- task test: 770 passed / 0 failed
- task bdd: 219 passed / 16 skip / 1 pre-existing budget_reset next_tick flake

+++ crates/aigw-migrate/src/native.rs
@@ -730,7 +730,7 @@ fn stream_pg_rows_keyset<'a>(
 
             // Clamp batch size to remaining limit when one is set.
             let limit_clause = match remaining {
-                Some(rem) if (rem as usize) < batch_size => rem as usize,
+                Some(rem) if rem < batch_size => rem,
                 _ => batch_size,
             };
             sql.push_str(&format!(" LIMIT {}", limit_clause));
@@ -883,7 +883,7 @@ fn stream_pg_rows_keyset_aigw<'a>(
 
             // Clamp batch size to remaining limit when one is set.
             let limit_clause = match remaining {
-                Some(rem) if (rem as usize) < batch_size => rem as usize,
+                Some(rem) if rem < batch_size => rem,
                 _ => batch_size,
             };
             sql.push_str(&format!(" LIMIT {}", limit_clause));
